requirements.txt
pandas
python-dotenv
polygon-api-client
requests